requirements.txt
Django==3.1.6
django-admin-view-permission==1.9
django-autocomplete-light==3.8.1
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-filebrowser-no-grappelli==3.8.0
django-simple-captcha==0.5.13
django-tinymce4-lite==1.8.0
django-utils-six==2.0
ipython
jsonfield==3.1.0
Pillow==8.1.0
pre-commit==2.10.0
sentry-sdk==0.19.5
URLObject==2.4.3
whitenoise==5.2.0
